Initial directory structure
This commit is just to create a base directory structure.

There are a couple of dummy remove.txt files, their purpose is to
be a placeholder until an actual testcase is added.

A dummy zuul job is added, on a subsequent commit, real zuul jobs
will be created to manage specific repository needs, such as
creating doc and release notes.

Change-Id: I232981ae3998607fcef7710ea33f8a5e76a8c601