Manual Tests: Networking LLDP Subdomain
With this patch, we are starting to upload LLDP test cases, the objective for
these test is to prove the LLDP agent functionallity. This agent is going to
show node properties according with the project needs, such as , interfaces,
OS and go on. Good number of the test cases still on development. We will
complement this domain with the rest of the test cases as soon as possible.

Change-Id: Ifa7cfadcd7daffe3fa15b0fb48520dcb75089a59