Manual Tests: Networking Internal DNS Resolution
This patch includes the first DNS test. The objective for this test is to
evaluate and demostrate how does the DNS service is going to affect our
communication with VM's. Following the logic, we can not reach any VM through
their name if the DNS services is not enabled. This first test is for AIO.

Change-Id: I38c5c0e53f1e80ae4c6506d39fd21163ebde9e05
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>