Manual Tests: Networking DPDK Subdomain
This patch contains the basic tests that we need in order to verify that our
StarlingX environment contains those DPDK/OVS versions that we want to work
with. First step in order to keep our environment should be to review our
desired packages. We are going to include more test cases related to this in
the future.

Change-Id: I9cafedc0ee27be61a152fd07549aa53e70aaf5a3