Update unlock on compute nodes
During the last week of february, there was a change on the apply
behavior for application platform-integ-apps.

After unlocking controller-1, it starts to do the apply of this
application.

For our standar configuration (2+2), this means that, when it tries
to do the unlock of the first compute node, the command fails because
the app was in the middle of the apply action.

This change adds a check, before doing a unlock of a compute node,
the suite will check and confirm that the application
platform-integ-apps is on applied status for 30 minutes. Until then,
it will try to execute the unlock command.

Change-Id: I7b346919e6ed6fb47e13c30a812bc262656aa4f5
Signed-off-by: George Postolache <george.postolache@intel.com>