Fix test cases related to lock/unlock of compute and storage nodes
This commit will prevent errors when trying to do the lock/unlock
on compute and storage nodes.

Right now, operation fails for compute-0 and oftenly on storage-0
because ceph monitor is not on sync.

This commit will change the Test Cases to be executed only for
compute-1 and storage-1. It'll also avoid increasing the execution
time.

Change-Id: If5c72b1cfd6bec2f74bfeae7dfcfd91324ef16a6
Signed-off-by: George Postolache <george.postolache@intel.com>