Add test_k8s_dashboard part2
a follow-up to #941156, adding an extra function(which is not finished
yet) to create a k8s dashboard, but it has the starting steps like
crating a secret and adding the https certificates, copying another
necessary files too.
Also, the abstraction for k8s secrets had to be improved by adding the
secret, table parser and output classes.
there was a typo in the kubectl_get_pods_output script which is now
fixed.

Change-Id: Ie6d60b893ec5b686e19dae8ea6197c3991e5b034
Signed-off-by: Gabriel Calixto de Paula <gabrielcalixto9@gmail.com>