PTP config value was added but is not in the default config file
A new value was added to the ptp config. This value was not added to
the default.json5 file that is loaded by default and crashing automation.
Added the value to the file.

Also fixed an issue exposed by this. If a test crashed quickly, we can have
a none start time. This will crash the upload to the db.

Change-Id: I936dfc614bbb52a2700404a2867f5ed01a925db8