Adding support for multiple port data settings
Updating the pmc parser to allow for multiple setting values. This
meant a refactor of other classes using this table. For now only
the port data seems to have multiple so other formats will fail if
more than one is found.

Also added some new unit tests for formats not covered previously
and fixed a couple of other tests.

Change-Id: I401ae0327399e23bcf2d899f048fe28228246f42