Subcloud backup restore test cases
This commit adds subcloud the following backup
restore test cases:

Restore a subcloud group with backup stored in
central cloud and subcloud are running an active
load.

Restore a subcloud group with backup stored in
the subclouds and subclouds are running an active
load.

Restore a subcloud backup with docker registry
images.

Change-Id: I6ec93c4751d412c1ef7f65bde194c634e5eb0b6e
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>