Backup and restore test cases
This commit introduces the following test cases
and its dependencies:

Backup creation for a subcloud with an inactive load,
triggered by controller-1 and stored on central cloud.

Backup creation for a subcloud with an inactive load,
triggered by controller-1 and stored in the subcloud.

Backup and restore from a subcloud running an active
load, triggered from controller-1. Backup is stored
on central cloud.

Backup failure due to lack of space in the default
backup directory ("/opt/dc-vault").

Backup deletion rejection due to subcloud failed
state.

Backup stored in central cloud with controller-1
as active controller and active load.

Backup stored in subcloud with controller-1
as active controller and active load.

Change-Id: I5d1fd899a60dc827e489fe86a39643ae025ae852
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>